+++ src/internal.h
@@ -225,7 +225,7 @@ struct _GLFWwindow
     // Window input state
     GLboolean           stickyKeys;
     GLboolean           stickyMouseButtons;
-    int                 cursorPosX, cursorPosY;
+    double              cursorPosX, cursorPosY;
     int                 cursorMode;
     char                mouseButton[GLFW_MOUSE_BUTTON_LAST + 1];
     char                key[GLFW_KEY_LAST + 1];
@@ -344,7 +344,7 @@ const char* _glfwPlatformGetVersionString(void);
 /*! @copydoc glfwSetCursorPos
  *  @ingroup platform
  */
-void _glfwPlatformSetCursorPos(_GLFWwindow* window, int xpos, int ypos);
+void _glfwPlatformSetCursorPos(_GLFWwindow* window, double xpos, double ypos);
 
 /*! @brief Sets up the specified cursor mode for the specified window.
  *  @param[in] window The window whose cursor mode to change.
@@ -614,7 +614,7 @@ void _glfwInputMouseClick(_GLFWwindow* window, int button, int action);
  *  of the client area of the window.
  *  @ingroup event
  */
-void _glfwInputCursorMotion(_GLFWwindow* window, int x, int y);
+void _glfwInputCursorMotion(_GLFWwindow* window, double x, double y);
 
 /*! @brief Notifies shared code of a cursor enter/leave event.
  *  @param[in] window The window that received the event.
